adding Element.Depth and Image.FlipEffect and Text.FlipEffect attributes

+++ Source/Lunohod/Lunohod/SpriteBatchWithFloats.cs
@@ -16,7 +16,7 @@ namespace Lunohod
         {
         }
 
-        internal void Draw(Texture2D texture2D, System.Drawing.RectangleF screenBounds, Rectangle? sourceRectangle, Color color, double rotation, Vector2 origin, SpriteEffects spriteEffects, int layerDepth)
+        internal void Draw(Texture2D texture2D, System.Drawing.RectangleF screenBounds, Rectangle? sourceRectangle, Color color, double rotation, Vector2 origin, SpriteEffects spriteEffects, float layerDepth)
         {
             screenBounds.ToRectangle(ref tmpRect);
